Handle missing part of an internal loop.
Sometimes only one half of an internal loop exists. In this case the correct
thing to do when asked for indicies with flanking pairs is to first extract the
flanking from the side that exists. Then use the closing pairs from that half to
get the pairs for the other side.