Added a basic RNAalifold wrapper.
RNAalifold can fold a sequence alignment unlike UNAfold and Mfold. This provides
a basic wrapper for the program. I have also changed the interface to .loops()
for parsers. It will now attempt to use self.sequence if no sequence is given.
This allows the results for RNAalifold to have a very similar interface to the
other result objects. In the case of RNAalifold it will set and use the conensus
sequence produced by RNAalifold.