Rename and complete first version.
I've renamed this since the ps file produced is actually from RNAplot so this
has been renamed to reflect that. In addition, I've added a first simple
version of the parser. It can load pairs, locations and sequence from a simple
file. The locations likely need to be converted from lower left to upper left
to be usable outside postscript, but I'll need to check.