Add a wrapper around PDB's PDBx package.
This is a wrapper around the PDBx parser to make it easier to deal with CIF
files. The parser they give is not very pythonic so I made this to make the
code easier to write. Plus there are some odd behaviors their code hos I'm
trying to cover up.