Add some methods to find polymers.
We can find polymers by looking for looking for things which are part of a
different asymmetric unit and then within each we look ones where the
auth_mon_id, as in the author assigned monomer is '?'. I think the '?' will
only occur when the unit was not observed in the structure.