Fix UNAfold failures
UNAfold was failing because I was not matching against the files correctly.
This corrects that. I've also added a case to test if there are any matched
files, if not it assumes folding failed somehow. I don't know how to trick the
program into doing that but it is a good check for the future.