Increase 2.5 compatibility
This is now sort of compatible with python 2.5. The one thing limiting this is
mutable sequence. I should change results set to not need this as it is just
for convenience. However, tox will still fail because I can't get it to install
the pdbx package from github which is needed for the cif file wrappers. I
should ask if pdb is going to release that as a package, which would make
things a lot easier.