Improved polymer detection
This way we find polymers by using the pdbx_unobs data which tells which
polymers are not observed. This is better than trying to infer it as I was
before. In addition, I've added several classes to represent things like models,
chains, and residues. This improves the design by allowing us to iterate over
parts of the structure to find polymers and generate unit ids.