Improve unit_ids module
The unit_ids module now contains code for generating and parsing three different
style of ids, Nucleotide, Unit and Matlab. I have written tests for the
Generators of all three. The parsers are still untested. In addition, I have
code for creating a converting function, as this is something simple that we do.