Add wrapper for SVG's produced by rnaplot
I now have a way to parse the svg files produced by RNAplot. This will prove
helpful when I finish the wrapper for RNAplot. In addition, I made changes to
make it easier to wrap RNAplot. That program does not read a file as input, but
instead reads from stdin. To deal with that I've altered the wrapper class to
support this. In addition, I renamed some of the methods for clarity and used
abc to make it clearer what has to be overridden to create a new wrapper.