Add code to map from experimental to resolved
Not all residues which went into the experiment are resolved. We need to know
the mapping between experimental and resolved for many different things. This
adds code to compute such a mapping using the pdbx_poly_seq_scheme table as
suggested by John. We now can get both the experimental sequence as well as a
mapping between a sequence id and the corresponding unit id. The sequence is
still in flux but this is a good start. We should also add some tests to make
sure the unit ids are generated correctly in all cases.