Allow for extraction of any size of flanking
Sometimes we want more than just the flanking pair, sometimes we want a whole
range. This is a first pass at such a functionality. We can now pass an integer
as well as a boolean to indices or loops to get the flanking. There are some
untested error cases, for example what if we get an internal loop where the
flanking count causes the extracted regions to overlap? This is not yet handled
but will work for simple uses.